Make invalid OutAddr states unrepresentable
This makes it impossible to represent wrong states for `OutAddr` (such
as a `NameAddr` + original `Name`). However, this requires changing
`OutAddr` to an enum similar to `Addr` but with an additional variant
for socket addresses with an original hostname. This introduces some
additional complexity, as the methods on `Addr` can no longer be used.

I'm not 100% sure if this is the right thing or not.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>